Working through getting the markdown files ready
Wrote up a basic markdown article on 'Wargaming' in the test.md. added some different parameters to see how markdown would react.

Speaking of React, Having issues, AGAIN with uploading the images, I'm pretty sure it's webpack that's the culprit but I managed to get the iamges up and running once before.. Imported a next-image loader to see if that will help.